pbr: Stop using TemporaryDirectory.
When using TemporaryDirectory in a with statement,
it wouldn't call the cleanup properly. So it would
literally wipe everything on the nfs mount point,
I got tired of dealing with it, so switched to manually
creating and removing the directory. Also added the
lazy umount option, for when it throws the device or
resource busy error.

Signed-off-by: Ryan Blakley <rblakley@redhat.com>