pbr: Add a check facts cmdline option.
The check facts option triggers comparing the temporarily written
facts files, to the ones under /var/lib/pbr/facts. If they differ
it returns 1, if they match it returns 0. This can be used in a
cron job, to determine if the iso needs to be rebuilt due to
storage, mount point, or cfg changes.

Signed-off-by: Ryan Blakley <rblakley@redhat.com>